feat: RadioButton context 사용 로직 추가, 위치 이동
